Reorganized source files for readability.
